scrummersApi/requirements.txt
django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.6.0
django-mysql==3.11.1
python-decouple==3.4